+++ README.md
@@ -41,7 +41,8 @@ Rename or copy `.env.example` to `.env` and fill it out. All variables are requi
 
 | Variable                      | Description                                                                                           |
 | ----------------------------- | ----------------------------------------------------------------------------------------------------- |
-| `TWITCH_USERNAME`             | The Twitch username to watch for streams                                                              |
+| `TWITCH_STREAMER_USERNAME`    | The Twitch username to watch for streams                                                              |
+| `TWITCH_BOT_USERNAME`         | The Twitch username for your chat bot                                                                 |
 | `TWITCH_CLIENT_ID`            | The client ID of your Twitch App                                                                      |
 | `TWITCH_CLIENT_SECRET`        | The client secret of your Twitch App                                                                  |
 | `TWITCH_EVENTSUB_HOSTNAME`    | The domain or public IP that your server can listen to for Twitch events                              |
@@ -49,6 +50,7 @@ Rename or copy `.env.example` to `.env` and fill it out. All variables are requi
 | `TWITCH_EVENTSUB_PORT`        | The port for your server to listen to for Twitch events. This is internal only, for the reverse proxy |
 | `TWITCH_BANNER_URL`           | _(optional)_ An image URL to use in stream notification embeds                                        |
 | `TWITCH_REDIRECT_URI`         | The URL that Twitch will send an authorization code to. Must be listed in your app's redirect URLs    |
+| `TWITCH_OAUTH_PORT`           | The port associated with your Twitch OAuth redirect URL. This must be set up in your reverse proxy    |
 | `TWITTER_USERNAME`            | The Twitter username to watch for tweets                                                              |
 | `TWITTER_TOKEN`               | The bearer token for your Twitter app                                                                 |
 | `TWITTER_INCLUDE_RETWEETS`    | If set to "true", retweets will be posted (quote retweets are always be posted)                       |
