+++ package.json
@@ -18,19 +18,24 @@
 	"author": "Devin Spikowski",
 	"license": "MIT",
 	"dependencies": {
-		"@twurple/api": "^5.3.3",
-		"@twurple/auth": "^5.3.3",
-		"@twurple/chat": "^5.3.3",
-		"@twurple/eventsub-http": "^5.3.3",
-		"@twurple/eventsub-ngrok": "^5.3.3",
+		"@twurple/api": "^6.0.5",
+		"@twurple/auth": "^6.0.5",
+		"@twurple/chat": "^6.0.5",
+		"@twurple/eventsub-http": "^6.0.5",
+		"@twurple/eventsub-ngrok": "^6.0.5",
 		"discord.js": "^14.7.1",
 		"dotenv": "^16.0.3",
+		"express": "^4.18.2",
+		"express-async-errors": "^3.1.1",
+		"express-session": "^1.17.3",
 		"lowdb": "^5.0.5",
 		"luxon": "^3.2.1",
 		"randomstring": "^1.2.3",
 		"twitter-api-v2": "^1.14.0"
 	},
 	"devDependencies": {
+		"@types/express": "^4.17.17",
+		"@types/express-session": "^1.17.6",
 		"@types/luxon": "^3.2.0",
 		"@types/node": "^18.11.18",
 		"@types/randomstring": "^1.1.8",
