+++ src/views/success.html
@@ -1,0 +1,46 @@
+<!DOCTYPE html>
+<html>
+	<head>
+		<title>Spice Bot</title>
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<meta charset="UTF-8" />
+		<style>
+			@import url('https://fonts.googleapis.com/css2?family=Roboto&display=swap');
+			html {
+				background: #221133;
+				color: #fef;
+				font-family: 'Roboto', sans-serif;
+			}
+			body {
+				display: flex;
+				flex-direction: column;
+				align-items: center;
+				padding: 1rem;
+			}
+			main {
+				background: #15081f;
+				padding: 0 2rem;
+				border-radius: 1rem;
+				max-width: 30rem;
+			}
+			h1 {
+				font-size: 3rem;
+			}
+			h2 {
+				text-align: center;
+			}
+			p {
+				margin: 1.5rem 0;
+				font-size: 1rem;
+				text-align: center;
+			}
+		</style>
+	</head>
+	<body>
+		<header><h1>🌶️🤖🤙</h1></header>
+		<main>
+			<h2>Spice Bot is ready to rock!</h2>
+			<p>You may now close this page</p>
+		</main>
+	</body>
+</html>
