+++ src/views/wrong-account.html
@@ -1,0 +1,68 @@
+<!DOCTYPE html>
+<html>
+	<head>
+		<title>Spice Bot</title>
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<meta charset="UTF-8" />
+		<style>
+			@import url('https://fonts.googleapis.com/css2?family=Roboto&display=swap');
+			html {
+				background: #221133;
+				color: #fef;
+				font-family: 'Roboto', sans-serif;
+			}
+			body {
+				display: flex;
+				flex-direction: column;
+				align-items: center;
+				padding: 1rem;
+			}
+			main {
+				background: #15081f;
+				padding: 0 2rem;
+				border-radius: 1rem;
+				max-width: 30rem;
+			}
+			h1 {
+				font-size: 3rem;
+			}
+			h2 {
+				text-align: center;
+			}
+			p {
+				margin: 1.5rem 0;
+				font-size: 1rem;
+				text-align: center;
+			}
+			strong {
+				color: #e735c1;
+			}
+			a {
+				display: inline-block;
+				color: #fef;
+				font-size: 1.25rem;
+				padding: 1rem;
+				background: #e735c1;
+				border: 3px solid #e735c1;
+				border-radius: 1rem;
+				font-weight: bold;
+				text-decoration: none;
+				transition: all 100ms ease-out;
+			}
+			a:hover,
+			a:active {
+				background: transparent;
+				border-color: #e735c1;
+				color: #e735c1;
+			}
+		</style>
+	</head>
+	<body>
+		<header><h1>Something isn't right...</h1></header>
+		<main>
+			<h2>You are using the wrong Twitch account</h2>
+			<p><a href="auth">Try Again</a></p>
+			<p>Make sure you are logged in as <strong>ybbaaabby</strong></p>
+		</main>
+	</body>
+</html>
